+++ android/src/main/kotlin/com/xzp/qrcode_flutter/QRCaptureViewFactory.kt
@@ -6,10 +6,10 @@ import io.flutter.plugin.common.StandardMessageCodec
 import io.flutter.plugin.platform.PlatformView
 import io.flutter.plugin.platform.PlatformViewFactory
 
-class QRCaptureViewFactory(private val registrar: PluginRegistry.Registrar) :
+class QRCaptureViewFactory() :
         PlatformViewFactory(StandardMessageCodec.INSTANCE) {
 
     override fun create(context: Context, id: Int, obj: Any?): PlatformView {
-        return QRCaptureView(registrar, id)
+        return QRCaptureView(id)
     }
 }